Move StopWatchy Package Code Shared Code
