+++ packages/ruby/shared-code/stopwatchy/stopwatchneverstartederror.rb
@@ -1,0 +1,3 @@
+class StopWatchNeverStartedError < StandardError
+
+end
