+++ packages/ruby/shared-code/stopwatchy/stopwatchneverstoppederror.rb
@@ -1,0 +1,3 @@
+class StopWatchNeverStoppedError < StandardError
+
+end
